+++ drivers/lora/CMakeLists.txt
@@ -12,5 +12,5 @@ endif()
 zephyr_library_sources_ifdef(CONFIG_LORA_SHELL shell.c)
 zephyr_library_sources_ifdef(CONFIG_HAS_SEMTECH_RADIO_DRIVERS hal_common.c)
 zephyr_library_sources_ifdef(CONFIG_HAS_SEMTECH_RADIO_DRIVERS sx12xx_common.c)
-zephyr_library_sources_ifdef(CONFIG_LORA_SX1276 sx1276.c)
+zephyr_library_sources_ifdef(CONFIG_LORA_SX127X sx127x.c)
 zephyr_library_sources_ifdef(CONFIG_LORA_SX126X sx126x.c)
