+++ drivers/lora/Kconfig.sx12xx
@@ -4,6 +4,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
+DT_COMPAT_SEMTECH_SX1272 := semtech,sx1272
 DT_COMPAT_SEMTECH_SX1276 := semtech,sx1276
 DT_COMPAT_SEMTECH_SX1261 := semtech,sx1261
 DT_COMPAT_SEMTECH_SX1262 := semtech,sx1262
@@ -20,7 +21,8 @@ if LORA_SX12XX
 
 choice
 	prompt "LoRa Radio chipset"
-	default LORA_SX1276 if $(dt_compat_enabled,$(DT_COMPAT_SEMTECH_SX1276))
+	default LORA_SX127X if $(dt_compat_enabled,$(DT_COMPAT_SEMTECH_SX1272))
+	default LORA_SX127X if $(dt_compat_enabled,$(DT_COMPAT_SEMTECH_SX1276))
 	default LORA_SX126X if $(dt_compat_enabled,$(DT_COMPAT_SEMTECH_SX1261))
 	default LORA_SX126X if $(dt_compat_enabled,$(DT_COMPAT_SEMTECH_SX1262))
 	help
@@ -28,11 +30,12 @@ choice
 	  is discovered from the device tree and should be correct for
 	  most users.
 
-config LORA_SX1276
-	bool "Semtech SX1276 driver"
-	select HAS_SEMTECH_SX1276
+config LORA_SX127X
+	bool "Semtech SX127x driver"
+	select HAS_SEMTECH_SX1272 if $(dt_compat_enabled,$(DT_COMPAT_SEMTECH_SX1272))
+	select HAS_SEMTECH_SX1276 if $(dt_compat_enabled,$(DT_COMPAT_SEMTECH_SX1276))
 	help
-	  Enable LoRa driver for Semtech SX1276.
+	  Enable LoRa driver for Semtech SX1272 and SX1276.
 
 config LORA_SX126X
 	bool "Semtech SX126x driver"
